Zmiana LONG na INT w całym projekcie

+++ wsb-jpa1/src/main/java/com/capgemini/wsb/persistence/dao/AddressDao.java
@@ -2,7 +2,7 @@ package com.capgemini.wsb.persistence.dao;
 
 import com.capgemini.wsb.persistence.entity.AddressEntity;
 
-public interface AddressDao extends Dao<AddressEntity, Long>
+public interface AddressDao extends Dao<AddressEntity, Integer>
 {
 
 }
