+++ wsb-jpa1/src/main/java/com/capgemini/wsb/persistence/dao/DoctorDao.java
@@ -4,7 +4,7 @@ import com.capgemini.wsb.persistence.entity.DoctorEntity;
 
 import java.util.List;
 
-public interface DoctorDao extends Dao<DoctorEntity, Long>
+public interface DoctorDao extends Dao<DoctorEntity, Integer>
 {
     List<DoctorEntity> findByDescription(final String pDesc);
 }
