+++ wsb-jpa1/src/main/java/com/capgemini/wsb/persistence/dao/MedicalTreatmentDao.java
@@ -2,7 +2,7 @@ package com.capgemini.wsb.persistence.dao;
 
 import com.capgemini.wsb.persistence.entity.MedicalTreatmentEntity;
 
-public interface MedicalTreatmentDao extends Dao<MedicalTreatmentEntity, Long>
+public interface MedicalTreatmentDao extends Dao<MedicalTreatmentEntity, Integer>
 {
 
 }
