+++ wsb-jpa1/src/main/java/com/capgemini/wsb/persistence/dao/VisitDao.java
@@ -2,7 +2,7 @@ package com.capgemini.wsb.persistence.dao;
 
 import com.capgemini.wsb.persistence.entity.VisitEntity;
 
-public interface VisitDao extends Dao<VisitEntity, Long>
+public interface VisitDao extends Dao<VisitEntity, Integer>
 {
 
 }
